Rename Midea LAN to midea
