+++ CODEOWNERS
@@ -1134,8 +1134,8 @@ CLAUDE.md @home-assistant/core
 /tests/components/metoffice/ @MrHarcombe @avee87
 /homeassistant/components/microbees/ @microBeesTech
 /tests/components/microbees/ @microBeesTech
-/homeassistant/components/midea_lan/ @chemelli74 @rokam @wuwentao
-/tests/components/midea_lan/ @chemelli74 @rokam @wuwentao
+/homeassistant/components/midea/ @chemelli74 @rokam @wuwentao
+/tests/components/midea/ @chemelli74 @rokam @wuwentao
 /homeassistant/components/miele/ @astrandb
 /tests/components/miele/ @astrandb
 /homeassistant/components/mikrotik/ @engrbm87 @chemelli74
