+++ homeassistant/components/midea/device_catalog.py
