+++ homeassistant/components/midea/entity.py
