+++ homeassistant/components/midea/manifest.json
@@ -1,9 +1,9 @@
 {
-  "domain": "midea_lan",
-  "name": "Midea LAN",
+  "domain": "midea",
+  "name": "Midea",
   "codeowners": ["@chemelli74", "@rokam", "@wuwentao"],
   "config_flow": true,
-  "documentation": "https://www.home-assistant.io/integrations/midea_lan",
+  "documentation": "https://www.home-assistant.io/integrations/midea",
   "integration_type": "device",
   "iot_class": "local_polling",
   "loggers": ["midealocal"],
