+++ homeassistant/components/midea/quality_scale.yaml
