+++ homeassistant/components/midea/strings.json
@@ -93,7 +93,7 @@
           "manually": "Configure manually",
           "search": "Search automatically"
         },
-        "title": "Set up Midea LAN"
+        "title": "Set up Midea"
       }
     }
   },
