+++ tests/components/midea/__init__.py
@@ -1,4 +1,4 @@
-"""Tests for the Midea LAN integration."""
+"""Tests for the Midea integration."""
 
 from unittest.mock import patch
 
@@ -12,11 +12,11 @@ from tests.common import MockConfigEntry
 async def setup_integration(
     hass: HomeAssistant, config_entry: MockConfigEntry, device: DummyDevice
 ) -> None:
-    """Set up a Midea LAN config entry backed by a fake device."""
+    """Set up a Midea config entry backed by a fake device."""
     config_entry.add_to_hass(hass)
 
     with patch(
-        "homeassistant.components.midea_lan.device_selector",
+        "homeassistant.components.midea.device_selector",
         return_value=device,
     ):
         assert await hass.config_entries.async_setup(config_entry.entry_id)
