+++ tests/components/midea/snapshots/test_climate.ambr
@@ -61,7 +61,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': None,
-    'platform': 'midea_lan',
+    'platform': 'midea',
     'previous_unique_id': None,
     'suggested_object_id': None,
     'supported_features': <ClimateEntityFeature: 441>,
@@ -161,7 +161,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': 'Zone 1 thermostat',
-    'platform': 'midea_lan',
+    'platform': 'midea',
     'previous_unique_id': None,
     'suggested_object_id': None,
     'supported_features': <ClimateEntityFeature: 385>,
@@ -233,7 +233,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': 'Zone 2 thermostat',
-    'platform': 'midea_lan',
+    'platform': 'midea',
     'previous_unique_id': None,
     'suggested_object_id': None,
     'supported_features': <ClimateEntityFeature: 385>,
@@ -322,7 +322,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': None,
-    'platform': 'midea_lan',
+    'platform': 'midea',
     'previous_unique_id': None,
     'suggested_object_id': None,
     'supported_features': <ClimateEntityFeature: 441>,
@@ -414,7 +414,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': None,
-    'platform': 'midea_lan',
+    'platform': 'midea',
     'previous_unique_id': None,
     'suggested_object_id': None,
     'supported_features': <ClimateEntityFeature: 385>,
@@ -495,7 +495,7 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': None,
-    'platform': 'midea_lan',
+    'platform': 'midea',
     'previous_unique_id': None,
     'suggested_object_id': None,
     'supported_features': <ClimateEntityFeature: 401>,
